With No Image

+++ app/Helpers/File-upload.php
@@ -8,7 +8,8 @@ function uploadFile(object $file, string $uploadPath, string $oldFile = null)
     $file_path = public_path($oldFile);
 
     if ($file_path) {
-        if (file_exists($oldFile)) {
+        if (file_exists($oldFile) && strpos($oldFile, noImage()) === false) {
+
             unlink($file_path);
         }
     }
@@ -29,7 +30,7 @@ function deleteFile(string $fileName)
 
     $file_path = public_path($fileName);
     // dd($file_path);
-    if ($file_path) {
+    if ($file_path && strpos($file_path, noImage() )  === false) {
         if (file_exists($file_path)) {
             unlink($file_path);
         }
